Tags selector, category and tag deletion fix
Tags and posts now have N:N relationship, categories and tags cannot be deleted if they are associated with a post to prevent possible application crashes

+++ app/Http/Controllers/CategoriesController.php
@@ -103,6 +103,11 @@ class CategoriesController extends Controller
     {
         $category = Category::findOrFail($id);
         $deleted_name = $category->name;
+        if($category->posts->count() > 0 ){
+            session()->flash('error', 'Category "'.$deleted_name.'" belongs to post(s), can\'t be deleted');
+
+            return redirect()->back();
+        }
         $category->delete();
 
         return redirect(route('categories.index'))->with('success', 'Category "'.$deleted_name.'" Deleted Successfully');
